+++ dtwin-client/src/components/onboarding/SplashScreen.jsx
