+++ dtwin-client/src/components/onboarding/StartingPage.jsx
